refactor: 重构 FunctionRegister 部分
