+++ src/main/java/com/ecnu/adsmls/views/codepage/CodePageController.java
@@ -2,10 +2,10 @@ package com.ecnu.adsmls.views.codepage;
 
 import com.alibaba.fastjson.JSON;
 import com.ecnu.adsmls.components.editor.Editor;
-import com.ecnu.adsmls.components.editor.modeleditor.LocationRegister;
+import com.ecnu.adsmls.utils.register.impl.LocationRegister;
 import com.ecnu.adsmls.components.editor.modeleditor.ModelEditor;
 import com.ecnu.adsmls.components.editor.treeeditor.TreeEditor;
-import com.ecnu.adsmls.components.editor.treeeditor.impl.BehaviorRegister;
+import com.ecnu.adsmls.utils.register.impl.BehaviorRegister;
 import com.ecnu.adsmls.components.modal.*;
 import com.ecnu.adsmls.components.mutileveldirectory.MultiLevelDirectory;
 import com.ecnu.adsmls.model.MConfig;
